remove `handled` state from calls and simplify the call state machine
Basically remove all the extraneous events that make no sense.

The purpose of the call state machine is to keep track of the state of a call so that it can be re-pooled if needed.

Whether a call is available or not for sending to an agent is no longer handled by the call state machine.
The information is now stored in the `handlers` counter associated with a call, so that we can properly decide for broadcast calls.